server/pyproject.toml
[tool.poetry]
name = "server"
version = "1.0.0"
description = "The Back-End of the portfolio website"
authors = ["Debbal Lotfi <debbal.lotfi.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
django-environ = "^0.11.2"
psycopg = {extras = ["binary"], version = "^3.2.1"}
djangorestframework = "^3.15.2"
markdown = "^3.6"
django-filter = "^24.2"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
